SmartInventorySystem/ai-service/requirements.txt
Flask
Flask-CORS
pandas
numpy
scikit-learn
statsmodels
pymongo
python-dotenv
joblib
